Second commit

+++ Product.java
@@ -17,10 +17,9 @@ public class Product {
         return name;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
-
     @Override
     public String toString() {
         return "Product{" +
